Enable Softwarecollections on RHEL and CentOS
As RHEL and CentOS are both longer-lived enterprise-class operating
systems the software versions are usually lagging a bit behind.

In case one wants the longevity of an Enterprise Linux together with
more modern software the Softwarecollections project exists which
offers amongst others more modern PHP releases.

This commit adds a pillar toggle to use the SCL repo and the ability
to select which php version to install.